Improve tracking of elapsed time
The elapsed time is now calculated per-request by comparing the current time with the gameStartTime. Therefore Game doesn't need a timer anymore. This is not only more effective, but also removes the need for the Game to implement the IDisposable interface.
Also fixed the gameStartTime initialization so the elapsed time is now calculated properly.
Added gameEndTime, so in case the game has already ended, ElapsedTime will return the time that has elapsed through the game.